SW_lab/sw_lab_part3/exercise3/requirements.txt
# Required
cherrypy == 18.6.0
paho-mqtt == 1.5.0
requests == 2.24.0

